fix(engine): 레벨 검출 네이티브 float 캐스트 + 병합 분기 테스트

+++ engine/flatness/core/levels.py
@@ -15,11 +15,11 @@ def detect_levels(median_z, bin_m=0.01, min_frac=0.03, merge_m=0.05):
         left = hist[i - 1] if i > 0 else 0
         right = hist[i + 1] if i < len(hist) - 1 else 0
         if hist[i] >= thresh and hist[i] >= left and hist[i] >= right:
-            peaks.append(0.5 * (edges[i] + edges[i + 1]))
+            peaks.append(float(0.5 * (edges[i] + edges[i + 1])))
     merged = []  # merge_m 이내로 붙은 피크는 하나의 레벨로 병합
     for p in peaks:
         if merged and p - merged[-1] < merge_m:
-            merged[-1] = 0.5 * (merged[-1] + p)
+            merged[-1] = float(0.5 * (merged[-1] + p))
         else:
             merged.append(p)
     return merged
